Remove primitive resolver
