+++ scarb/src/resolver/in_memory_index.rs
