--- scarb/src/resolver/primitive.rs
@@ -1,209 +1,0 @@
-use crate::core::lockfile::Lockfile;
-use crate::core::registry::Registry;
-use crate::core::resolver::DependencyEdge;
-use crate::core::{
-    DepKind, DependencyFilter, DependencyVersionReq, ManifestDependency, PackageId, Resolve,
-    Summary,
-};
-use anyhow::{Result, bail};
-use indoc::{formatdoc, indoc};
-use petgraph::graphmap::DiGraphMap;
-use std::collections::{HashMap, HashSet};
-
-/// Builds the list of all packages required to build the first argument.
-///
-/// This is meant to be used by `scarb::resolver::resolve` function.
-///
-/// # Arguments
-///
-/// * `summaries` - the list of all top-level packages that are intended to be part of
-///     the lock file (resolve output).
-///     These typically are a list of all workspace members.
-///
-/// * `registry` - this is the source from which all package summaries are loaded.
-///     It is expected that this is extensively configured ahead of time and is idempotent with
-///     our requests to it (aka returns the same results for the same query every time).
-///     It is also advised to implement internal caching, as the resolver may frequently ask
-///     repetitive queries.
-///
-/// * `lockfile` - a [`Lockfile`] instance, which is used to guide the resolution process. Empty
-///     lockfile will result in no guidance. This function does not read or write lock files from
-///     the filesystem.
-///
-/// # Implementation:
-///
-/// This is a naive implementation of the dependency resolver.
-/// It will choose the first available version of a dependency that satisfies the version requirement.
-/// It will not attempt to solve any package incompatibilities and return an error instead.
-#[tracing::instrument(level = "trace", skip_all)]
-pub async fn resolve(
-    summaries: &[Summary],
-    registry: &dyn Registry,
-    lockfile: Lockfile,
-) -> Result<Resolve> {
-    let mut graph = DiGraphMap::<PackageId, DependencyEdge>::new();
-
-    let main_packages = summaries
-        .iter()
-        .map(|sum| sum.package_id)
-        .collect::<HashSet<PackageId>>();
-    let mut packages: HashMap<_, _> = HashMap::from_iter(
-        summaries
-            .iter()
-            .map(|s| (s.package_id.name.clone(), s.package_id)),
-    );
-
-    let mut summaries: HashMap<_, _> = summaries
-        .iter()
-        .map(|s| (s.package_id, s.clone()))
-        .collect();
-
-    let mut queue: Vec<PackageId> = summaries.keys().copied().collect();
-    while !queue.is_empty() {
-        let mut next_queue = Vec::new();
-
-        for package_id in queue {
-            graph.add_node(package_id);
-
-            let summary = summaries[&package_id].clone();
-            let dep_filter =
-                DependencyFilter::propagation(main_packages.contains(&summary.package_id));
-            for dep in summary.filtered_full_dependencies(dep_filter) {
-                let dep = rewrite_dependency_source_id(registry, &package_id, dep).await?;
-
-                let locked_package_id = lockfile.package_matching(dep.clone());
-                let dep = if let Some(locked_package_id) = locked_package_id {
-                    rewrite_locked_dependency(dep.clone(), locked_package_id?)
-                } else {
-                    dep
-                };
-
-                let results = registry.query(&dep).await?;
-
-                let Some(dep_summary) = results.first() else {
-                    bail!("cannot find package {}", dep.name)
-                };
-
-                let dep_kind = dep.kind.clone();
-                let dep = dep_summary.package_id;
-
-                if let Some(existing) = packages.get(dep.name.as_ref()) {
-                    if existing.source_id != dep.source_id {
-                        bail!(
-                            indoc! {"
-                            found dependencies on the same package `{}` coming from incompatible \
-                            sources:
-                            source 1: {}
-                            source 2: {}
-                            "},
-                            dep.name,
-                            existing.source_id,
-                            dep.source_id
-                        );
-                    }
-                }
-
-                let weight = graph.edge_weight_mut(summary.package_id, dep);
-                if let Some(weight) = weight {
-                    match dep_kind {
-                        DepKind::Normal => {
-                            weight.accept_all();
-                        }
-                        DepKind::Target(target_kind) => {
-                            weight.accept_new(target_kind);
-                        }
-                    };
-                } else {
-                    let weight = match dep_kind {
-                        DepKind::Normal => DependencyEdge::for_all_targets(),
-                        DepKind::Target(target_kind) => DependencyEdge::for_target(target_kind),
-                    };
-                    graph.add_edge(summary.package_id, dep, weight);
-                }
-
-                summaries.insert(dep, dep_summary.clone());
-
-                if packages.contains_key(dep.name.as_ref()) {
-                    continue;
-                }
-
-                packages.insert(dep.name.clone(), dep);
-                next_queue.push(dep);
-            }
-        }
-
-        queue = next_queue;
-    }
-
-    // Detect incompatibilities and bail in case ones are found.
-    let mut incompatibilities = Vec::new();
-    for from_package in graph.nodes() {
-        let dep_filter = DependencyFilter::propagation(main_packages.contains(&from_package));
-        for manifest_dependency in summaries[&from_package].filtered_full_dependencies(dep_filter) {
-            let to_package = packages[&manifest_dependency.name];
-            if !manifest_dependency.matches_package_id(to_package) {
-                let message = format!(
-                    "- {from_package} cannot use {to_package}, because {} requires {} {}",
-                    from_package.name, to_package.name, manifest_dependency.version_req
-                );
-                incompatibilities.push(message);
-            }
-        }
-    }
-
-    if !incompatibilities.is_empty() {
-        incompatibilities.sort();
-        let incompatibilities = incompatibilities.join("\n");
-        bail!(formatdoc! {"
-            Version solving failed:
-            {incompatibilities}
-
-            Scarb does not have real version solving algorithm yet.
-            Perhaps in the future this conflict could be resolved, but currently,
-            please upgrade your dependencies to use latest versions of their dependencies.
-        "});
-    }
-
-    let resolve = Resolve { graph, summaries };
-    resolve.check_checksums(&lockfile)?;
-    Ok(resolve)
-}
-
-fn rewrite_locked_dependency(
-    dependency: ManifestDependency,
-    locked_package_id: PackageId,
-) -> ManifestDependency {
-    ManifestDependency::builder()
-        .kind(dependency.kind.clone())
-        .name(dependency.name.clone())
-        .source_id(locked_package_id.source_id)
-        .version_req(DependencyVersionReq::Locked {
-            exact: locked_package_id.version.clone(),
-            req: dependency.version_req.clone().into(),
-        })
-        .build()
-}
-
-async fn rewrite_dependency_source_id(
-    registry: &dyn Registry,
-    package_id: &PackageId,
-    dependency: &ManifestDependency,
-) -> Result<ManifestDependency> {
-    // Rewrite path dependencies for git sources.
-    if package_id.source_id.is_git() && dependency.source_id.is_path() {
-        let rewritten_dep = ManifestDependency::builder()
-            .kind(dependency.kind.clone())
-            .name(dependency.name.clone())
-            .source_id(package_id.source_id)
-            .version_req(dependency.version_req.clone())
-            .build();
-        // Check if this dependency can be queried from git source.
-        // E.g. packages below other package's manifest will not be accessible.
-        if !registry.query(&rewritten_dep).await?.is_empty() {
-            // If it is, return rewritten dependency.
-            return Ok(rewritten_dep);
-        }
-    };
-
-    Ok(dependency.clone())
-}
